Rename the routes file
