+++ src/components/pages/home/CommonCategoryCard.tsx
@@ -1,5 +1,5 @@
 import { homeBrowseByCategory } from "@/mock/category";
-import { Routes } from "@/utils/constants/linksPaths";
+import { Routes } from "@/utils/constants/routes";
 import { Placeholder } from "@phosphor-icons/react";
 import Link from "next/link";
